+++ NEWS.md
@@ -1,7 +1,8 @@
 # tern 0.7.8.9007
 
 ### Migration from `assertthat` to `checkmate`
-* removed `is_quantiles_vector`.
+* removed `is_quantiles_vector` and `all_elements_in_ref` (substituted by 
+  `checkmate::subset()`).
 * removed `is_proportion_vector` and adapted `is_proportion` to `checkmate` style 
   with `assert_proportion_value`.
 * `is_equal_length` is now `assert_equal_length`.